baselines

+++ experiments/robot/openvla_utils.py
@@ -93,9 +93,9 @@ def hf_to_vllm(vla, processor, cfg):
     # Save language model 
     vllm_model_path = f"logs/{cfg.pretrained_checkpoint.replace('/', '_')}-vllm"
     
-    # if not os.path.exists(vllm_model_path):
-    vla.language_model.save_pretrained(vllm_model_path)
-    processor.save_pretrained(vllm_model_path)
+    if not os.path.exists(vllm_model_path):
+        vla.language_model.save_pretrained(vllm_model_path)
+        processor.save_pretrained(vllm_model_path)
 
     # Load language model with VLLM
     if hasattr(vla, "language_model"):
@@ -103,14 +103,26 @@ def hf_to_vllm(vla, processor, cfg):
     # TODO: check vllm load mode, check settings, memory
     # check if async engine is enabled
     if not hasattr(cfg, 'async_engine') or not cfg.async_engine:
-        vla.language_model = vllm.LLM(vllm_model_path, 
-                                      trust_remote_code=True, 
-                                      gpu_memory_utilization=0.7, 
-                                      preemption_mode='swap', 
-                                      swap_space = 10, 
-                                      enable_chunked_prefill = True, 
-                                      enable_prefix_caching = True, 
-                                      )
+        if not hasattr(cfg, 'quantization'):
+            vla.language_model = vllm.LLM(vllm_model_path, 
+                                        trust_remote_code=True, 
+                                        gpu_memory_utilization=0.7, 
+                                        preemption_mode='swap', 
+                                        swap_space = 10, 
+                                        enable_chunked_prefill = True, 
+                                        enable_prefix_caching = True, 
+                                        )
+        else:
+            vla.language_model = vllm.LLM(vllm_model_path, 
+                                        trust_remote_code=True, 
+                                        gpu_memory_utilization=0.7, 
+                                        preemption_mode='swap', 
+                                        swap_space = 10, 
+                                        enable_chunked_prefill = True, 
+                                        enable_prefix_caching = True, 
+                                        quantization="bitsandbytes", 
+                                        load_format="bitsandbytes"
+                                        )
     else:
         vla.language_model  = vllm.AsyncLLMEngine.from_engine_args(
                 vllm.AsyncEngineArgs(
@@ -223,9 +235,9 @@ def get_vla_action(vla, processor, base_vla_name, obs, task_label, unnorm_key, c
         # 3. VLLM inference batched 
         if hasattr(vla, 'use_vllm') and vla.use_vllm:
             import vllm # only executed once    
-            
-            if prompts is None: 
-                prompts = [prompt]
+
+            if prompts is None: prompts = [prompt]
+            if len(prompts) == 1:  
                 sampling_params = vllm.SamplingParams(temperature=0, max_tokens=max_new_tokens, stop_token_ids=[2]) # [EOS]
             else:
                 sampling_params = vllm.SamplingParams(temperature=0, max_tokens=max_new_tokens, stop_token_ids=[2, 29901]) # [EOS] and :
@@ -327,10 +339,11 @@ def get_vla_action_async(vla, processor, base_vla_name, obs, task_label, unnorm_
     if hasattr(vla, 'use_vllm') and vla.use_vllm:
         import vllm # only executed once
         with torch.no_grad():
-            if prompts is None: 
-                prompts = [prompt]
+            if prompts is None: prompts = [prompt]
+            if len(prompts) == 1:  
                 sampling_params = vllm.SamplingParams(temperature=0, max_tokens=max_new_tokens, stop_token_ids=[2]) # [EOS]
-                prefill = True
+            elif len(prompts) == 2:  
+                sampling_params = vllm.SamplingParams(temperature=0, max_tokens=max_new_tokens, stop_token_ids=[2]) #baseline
             else:
                 sampling_params = vllm.SamplingParams(temperature=0, max_tokens=max_new_tokens, stop_token_ids=[2, 29901]) # [EOS] and :
             prompts_reason = prompts[:-1]
@@ -382,13 +395,14 @@ def get_vla_action_async(vla, processor, base_vla_name, obs, task_label, unnorm_
 class CotTag(enum.Enum):
     TASK = "TASK:"
     PLAN = "PLAN:"
-    VISIBLE_OBJECTS = "VISIBLE OBJECTS:"
     SUBTASK_REASONING = "SUBTASK REASONING:"
     SUBTASK = "SUBTASK:"
     MOVE_REASONING = "MOVE REASONING:"
     MOVE = "MOVE:"
     GRIPPER_POSITION = "GRIPPER POSITION:"
+    VISIBLE_OBJECTS = "VISIBLE OBJECTS:"
     ACTION = "ACTION:"
+    
 
 class PromptManager(object):
                         
@@ -466,4 +480,79 @@ class PromptManager(object):
             new_prompt = prompt + "Action: "
             prompts.append(new_prompt)
             prompt = prompt + t.value + self.subtask_history[t.name][-1] # Use updated history
+        return prompts
+
+
+class PromptManager5step(object):
+                        
+    def __init__(self, history_adaptive=False):
+        # Intialize subtask history
+        self.history_adaptive = history_adaptive
+        self.history_idx = len(CotTag) - 1 
+        self.start_key = "TASK:"
+        self.end_key = "MOVE REASONING:"
+        self.subtask_history = ['']
+        self.highlevel_frequency = 5
+        self.update_counter = self.highlevel_frequency -1
+
+    def update_history(self, generated_text):
+        """ Update subtask history based on 
+            Args:
+                - index: if index is specified, only extract that subtask
+        """
+        if self.update_counter != self.highlevel_frequency - 1:
+            return
+        start_idx = generated_text.find(self.start_key)
+        end_idx = generated_text.find(self.end_key)
+        subtask_text =  generated_text[start_idx+len(self.start_key):end_idx]
+        if subtask_text != self.subtask_history[-1]:
+            self.subtask_history.append(subtask_text)
+            print(f"\033[91m subtext {self.start_key}{subtask_text}\033[0m") 
+
+    def generate_prompts(self, task_description):
+        """ Generate batch prompts, with history
+        """
+        prompts = []
+        prompt = f"{OPENVLA_V01_SYSTEM_PROMPT} USER: What action should the robot take to {task_description.lower()}? ASSISTANT: "
+        if self.update_counter != 0:
+            prompt = prompt + self.start_key + self.subtask_history[-1] + self.end_key # Use updated history
+            self.update_counter = self.update_counter - 1
+        else:
+            prompt = prompt + self.start_key 
+            self.update_counter = self.highlevel_frequency - 1
+        prompts.append(prompt)
+        # print(f"\033[93mprompt: {prompt}\033[0m")
+        # print(f"\033[91m Promts Length {len(prompts)}\033[0m")
+        return prompts
+    
+class PromptManagerAsyncBase(object):
+                        
+    def __init__(self, history_adaptive=False):
+        # Intialize subtask history
+        self.history_adaptive = history_adaptive
+        self.history_idx = len(CotTag) - 1 
+        self.start_key = "TASK:"
+        self.end_key = "MOVE REASONING:"
+        self.subtask_history = ['']
+
+    def update_history(self, generated_text):
+        """ Update subtask history based on 
+            Args:
+                - index: if index is specified, only extract that subtask
+        """
+        start_idx = generated_text.find(self.start_key)
+        end_idx = generated_text.find(self.end_key)
+        subtask_text =  generated_text[start_idx+len(self.start_key):end_idx]
+        if subtask_text != self.subtask_history[-1]:
+            self.subtask_history.append(subtask_text)
+            print(f"\033[91m subtext {self.start_key}{subtask_text}\033[0m") 
+
+    def generate_prompts(self, task_description):
+        """ Generate batch prompts, with history
+        """
+        prompts = []
+        prompt = f"{OPENVLA_V01_SYSTEM_PROMPT} USER: What action should the robot take to {task_description.lower()}? ASSISTANT: "
+        prompts.append(prompt + self.start_key) 
+        prompts.append(prompt + self.start_key + self.subtask_history[-1] + self.end_key) # Use updated history
+        print(f"\033[93mprompt: {prompts}\033[0m")
         return prompts
